Split genericmiot into parts for easier testing
